refactor(desktop): split registerIpc into per-domain register*Ipc modules
Migrate the un-migrated main.ts registerIpc() domains into dedicated
register*Ipc(deps) modules following the existing *-ipc-main.ts pattern,
so registerIpc becomes pure composition and each domain is testable in
isolation instead of only via module-side-effect source greps.

New modules: app-ipc-main (window + app info/openPath/project-select +
visualSmoke), git-ipc-main, workspace-search-ipc-main,
workspace-instructions-ipc-main, onboarding-ipc-main,
session-entry-ipc-main (quickChat + expertTeam), sessions-ipc-main,
permissions-ipc-main (permissions + capabilities + health),
settings-ipc-main (settings/bots/network), gateway-ipc-main. The shared
"current project root" selection moves into project-root-controller so a
single authority owns the state now that its readers and the picker
handlers live in separate modules.

Pure behavior- and visually-neutral move: handler bodies and shared
wiring (streamEvents, emitSessionsChanged, ensureSessionCanSend,
getReadyConnection, applySettingsRuntimeEffects, ...) are unchanged and
injected as deps. main.ts drops from 2447 to 1590 lines.

Co-migrate the source-reading contract tests: the wiring contract now
enumerates every extracted registrar, the combined-source helper lists
the new modules, and the direct main.ts readers switch to the combined
source. Allow-list the relocated visual-smoke capture log.

Refs #1084
