+++ apps/desktop/src/main/__tests__/main-process-contract-source-helpers.ts
@@ -6,6 +6,7 @@ const REPO_ROOT = resolve(import.meta.dirname, '../../../../..');
 
 export const MAIN_PROCESS_SOURCE_REPO_PATHS: readonly string[] = [
   'apps/desktop/src/main/main.ts',
+  'apps/desktop/src/main/app-ipc-main.ts',
   'apps/desktop/src/main/bot-incoming-main.ts',
   'apps/desktop/src/main/browser-ipc-main.ts',
   'apps/desktop/src/main/config-ipc-main.ts',
@@ -13,19 +14,29 @@ export const MAIN_PROCESS_SOURCE_REPO_PATHS: readonly string[] = [
   'apps/desktop/src/main/connections-ipc-main.ts',
   'apps/desktop/src/main/daily-review-ipc-main.ts',
   'apps/desktop/src/main/daily-review-main.ts',
+  'apps/desktop/src/main/gateway-ipc-main.ts',
+  'apps/desktop/src/main/git-ipc-main.ts',
   'apps/desktop/src/main/main-window.ts',
   'apps/desktop/src/main/memory-ipc-main.ts',
   'apps/desktop/src/main/notifications-ipc-main.ts',
   'apps/desktop/src/main/oauth-model-connections-main.ts',
+  'apps/desktop/src/main/onboarding-ipc-main.ts',
   'apps/desktop/src/main/permission-mode-default.ts',
+  'apps/desktop/src/main/permissions-ipc-main.ts',
   'apps/desktop/src/main/plan-reminders-ipc-main.ts',
   'apps/desktop/src/main/plan-reminders-main.ts',
+  'apps/desktop/src/main/project-root-controller.ts',
+  'apps/desktop/src/main/session-entry-ipc-main.ts',
+  'apps/desktop/src/main/sessions-ipc-main.ts',
+  'apps/desktop/src/main/settings-ipc-main.ts',
   'apps/desktop/src/main/subscription-model-fetch.ts',
   'apps/desktop/src/main/subscription-ipc-main.ts',
   'apps/desktop/src/main/system-prompt-main.ts',
   'apps/desktop/src/main/usage-ipc-main.ts',
   'apps/desktop/src/main/web-search-ipc-main.ts',
+  'apps/desktop/src/main/workspace-instructions-ipc-main.ts',
   'apps/desktop/src/main/workspace-resources-ipc-main.ts',
+  'apps/desktop/src/main/workspace-search-ipc-main.ts',
 ];
 
 export async function readMainProcessCombinedSource(): Promise<string> {
