+++ apps/desktop/src/main/__tests__/session-thinking-level-main-contract.test.ts
@@ -2,11 +2,12 @@ import { strict as assert } from 'node:assert';
 import { readFile } from 'node:fs/promises';
 import { resolve } from 'node:path';
 import { describe, it } from 'node:test';
+import { readMainProcessCombinedSource } from './main-process-contract-source-helpers.js';
 
 const REPO_ROOT = resolve(import.meta.dirname, '../../../../..');
 
 async function readMainSource(): Promise<string> {
-  return readFile(resolve(REPO_ROOT, 'apps/desktop/src/main/main.ts'), 'utf8');
+  return readMainProcessCombinedSource();
 }
 
 function extractIpcHandler(source: string, channel: string): string {
