+++ apps/desktop/src/main/main.ts
@@ -1,44 +1,23 @@
-import { app, ipcMain, nativeImage, powerMonitor, safeStorage, shell } from 'electron';
+import { app, nativeImage, powerMonitor, safeStorage, shell } from 'electron';
 import { randomUUID } from 'node:crypto';
-import { mkdir, readFile, realpath, stat, writeFile } from 'node:fs/promises';
-import { basename, isAbsolute, join, relative, resolve, sep } from 'node:path';
+import { mkdir, readFile, realpath } from 'node:fs/promises';
+import { isAbsolute, join, relative, resolve, sep } from 'node:path';
 import { startConfigFileWatcher, type ConfigFileWatcher } from './config-file-watcher.js';
-import { release as osRelease, arch as osArch } from 'node:os';
 import {
-  generalizedErrorMessage,
-  generalizedErrorMessageChinese,
-  redactSecrets,
   filterModelVisibleTaskLedgerTasks,
-  sanitizeTaskLedgerTask,
-  buildHealthSnapshot,
-  healthSignalFromCapability,
-  healthSignalFromConnection,
-  healthSignalFromConnectionRuntime,
-  isPermissionMode,
   isDeepResearchSession,
-  isThinkingLevel,
-  thinkingVariantsForModel,
   resolveModelVisionSupport,
   DEEP_RESEARCH_SESSION_LABEL,
   expertTeamIdFromLabels,
-  expertTeamLabel,
-  botDisplayLabel,
 } from '@maka/core';
 import type {
   AppSettings,
   BotProvider,
   ConnectionEvent,
-  CreateSessionInput,
   PermissionMode,
   SessionChangedEvent,
   SessionChangedReason,
   SessionEvent,
-  SessionHeader,
-  SessionListFilter,
-  ThinkingLevel,
-  StoredMessage,
-  SettingsTestResult,
-  UpdateAppSettingsResult,
   UpdateAppSettingsInput,
 } from '@maka/core';
 import { deriveBotStatusPersistenceUpdate } from './bot-status-persistence.js';
@@ -49,14 +28,6 @@ import {
   persistArchivedToolResultToArtifacts,
   readArchivedToolResultFromArtifacts,
 } from './tool-result-archive-artifacts.js';
-import {
-  normalizeBranchFromTurnInput,
-  normalizePermissionResponse,
-  normalizeRegenerateTurnInput,
-  normalizeSessionSendCommand,
-  normalizeStopSessionInput,
-  normalizeUserQuestionResponse,
-} from './permission-response-guard.js';
 import { turnFailureMessageFromSessionEvent } from './turn-stream-outcome.js';
 import { ClaudeSubscriptionService } from './oauth/claude-subscription-service.js';
 import { OpenAiCodexService } from './oauth/openai-codex-service.js';
@@ -65,13 +36,8 @@ import { CursorSubscriptionService } from './oauth/cursor-subscription-service.j
 import { AntigravitySubscriptionService } from './oauth/antigravity-subscription-service.js';
 import { importLegacyOAuthTokenFiles } from './oauth/shared-credential-bridge.js';
 import type { WorkspacePrivacyContext } from '@maka/core/incognito';
-import type { LlmCallRecord, PricingConfig, ToolInvocationRecord } from '@maka/core/usage-stats/types';
-import type {
-  TestProxyInput,
-  TestProxyResult,
-} from '@maka/core/settings/network-settings';
-import { SENSITIVE_PLACEHOLDER } from '@maka/core/settings/network-settings';
-import { err, ok, tryResult, type Result } from '@maka/core/settings/result';
+import type { LlmCallRecord, ToolInvocationRecord } from '@maka/core/usage-stats/types';
+import { ok } from '@maka/core/settings/result';
 import {
   AiSdkBackend,
   BackendRegistry,
@@ -89,15 +55,11 @@ import {
   buildSubagentSpawnTool,
   buildSubagentToolGroup,
   getAIModel,
-  getExpertTeam,
-  listExpertTeams,
   buildProviderOptions,
   recordLlmCall,
   recordToolInvocation,
   buildPricingLookup,
   BotRegistry,
-  getWechatBridgeQrCode,
-  testBotChannel as testRuntimeBotChannel,
   setActiveProxy,
   ShellRunProcessManager,
 } from '@maka/runtime';
@@ -111,9 +73,7 @@ import type {
   ToolResultArchiveReadResult,
   ToolResultArchiveRecorderInput,
 } from '@maka/runtime';
-import { testProxyConnection } from '@maka/runtime/network/proxy-test';
-import { fetchWeChatQrcode, pollWeChatQrcodeStatus } from './wechat-scan-login.js';
-import type { LlmConnection, ProviderType } from '@maka/core/llm-connections';
+import type { LlmConnection } from '@maka/core/llm-connections';
 import {
   createAgentRunStore,
   createArtifactStore,
@@ -132,35 +92,17 @@ import {
   errorReason,
   requireReadyConnection,
 } from './chat-readiness.js';
-import {
-  sessionReadMessagesFailureMessage,
-} from './session-read-error-copy.js';
 import { createFileCredentialStore, migrateLegacyCredentials } from './credential-store.js';
 import { bindOnboardingDeps, createOnboardingService } from './onboarding-service.js';
 import { handleQuickChatStart as runQuickChatStart, type QuickChatResult } from './quick-chat.js';
-import { handleExpertTeamStart as runExpertTeamStart } from './expert-team-start.js';
-import { probeOfficeCli } from './officecli-probe.js';
-import { resolveOpenPath, type OpenPathResult } from './open-path-guard.js';
-import { resolveProjectGitInfo, resolveProjectRoot, resolveSkillDiscoveryPaths } from '@maka/runtime';
-import { listLocalBranches, checkoutBranch } from './git-branch.js';
-import { searchWorkspaceFiles } from './workspace-file-search.js';
+import { resolveSkillDiscoveryPaths } from '@maka/runtime';
 import { createDailyReviewArchiveStore } from './daily-review-archive-store.js';
-import { botTestErrorMessage, buildSettingsUpdateResult, maskAppSettings, preserveSensitivePlaceholders, toSettingsTestResult } from './settings-ipc-helpers.js';
+import { preserveSensitivePlaceholders } from './settings-ipc-helpers.js';
 import {
   buildSkillAgentTool,
 } from './skills.js';
-import {
-  createWorkspaceInstructionFile,
-  getWorkspaceInstructionsState,
-  resolveWorkspaceInstructionFileForOpen,
-  type WorkspaceInstructionCreateFailureReason,
-  type WorkspaceInstructionOpenFailureReason,
-} from './workspace-instructions.js';
-import { buildCapabilitySnapshotCollection, buildPermissionSnapshot } from './capability-snapshot.js';
-import { openSystemPermissionPane, requestPermissionAccess } from './permissions-actions.js';
 import { resolveDefaultPermissionMode } from './permission-mode-default.js';
 import {
-  getVisualSmokeState,
   resolveVisualSmokeFixture,
   seedVisualSmokeFixture,
 } from './visual-smoke-fixture.js';
@@ -169,8 +111,6 @@ import { OpenGatewayService } from './open-gateway.js';
 import { LocalMemoryService } from './local-memory-service.js';
 import { createAttachmentApprovalRegistry } from './attachment-approval.js';
 import { createAttachmentByteReader } from './attachment-reader.js';
-import { resizeImageForAttachment } from './attachment-resize-native.js';
-import { resolveSessionSend } from './session-send-resolve.js';
 import { buildExploreAgentTool } from './explore-agent-tool.js';
 import { buildOfficeDocumentEditTool, buildOfficeDocumentTool } from './office-document-tool.js';
 import {
@@ -195,7 +135,6 @@ import {
   computerUseToolsForModel,
 } from './computer-use-model-tools.js';
 import { createComputerUseOverlayHook } from '@maka/computer-use';
-import { releaseBrowserSession } from './browser/session.js';
 import { createMainWindowController } from './main-window.js';
 import { createDailyReviewMainService } from './daily-review-main.js';
 import { createPlanReminderMainService } from './plan-reminders-main.js';
@@ -209,9 +148,7 @@ import { createMainGoalWiring } from './goal-wiring.js';
 import { handleGoalContinuation } from '@maka/runtime';
 import { createOAuthModelConnectionsMainService } from './oauth-model-connections-main.js';
 import {
-  applyNetworkPatch,
   maskNetworkSettings,
-  toAppNetworkPatch,
   toContractNetworkSettings,
 } from './network-settings-main.js';
 import { registerMemoryIpc } from './memory-ipc-main.js';
@@ -225,6 +162,17 @@ import { registerDailyReviewIpc } from './daily-review-ipc-main.js';
 import { registerUsageIpc } from './usage-ipc-main.js';
 import { registerWebSearchIpc } from './web-search-ipc-main.js';
 import { registerNotificationsIpc } from './notifications-ipc-main.js';
+import { registerAppIpc } from './app-ipc-main.js';
+import { registerGitIpc } from './git-ipc-main.js';
+import { registerWorkspaceSearchIpc } from './workspace-search-ipc-main.js';
+import { registerWorkspaceInstructionsIpc } from './workspace-instructions-ipc-main.js';
+import { registerOnboardingIpc } from './onboarding-ipc-main.js';
+import { registerSessionEntryIpc } from './session-entry-ipc-main.js';
+import { registerPermissionsIpc } from './permissions-ipc-main.js';
+import { registerSettingsIpc } from './settings-ipc-main.js';
+import { registerGatewayIpc } from './gateway-ipc-main.js';
+import { registerSessionsIpc } from './sessions-ipc-main.js';
+import { createProjectRootController } from './project-root-controller.js';
 
 // E2E switches must never fire in a packaged build, and must never run against
 // the real user data: a stray MAKA_E2E on a build/dev machine would otherwise
@@ -759,11 +707,15 @@ let lookupPricing = buildPricingLookup();
 // newer, more useful failure replaces the previous one.
 const previousBotStatus = new Map<BotProvider, Pick<BotStatus, 'readiness' | 'reason'>>();
 let botIncoming: ReturnType<typeof createBotIncomingMainService>;
-// botIncoming is wired at module load, before registerIpc() defines the
-// current-project-root resolver. registerIpc reassigns this once the resolver
-// exists; until then the launch directory is the safe fallback. Unifying
-// project-root resolution is tracked as a follow-up.
-let resolveCurrentProjectRoot: () => Promise<string> = async () => process.cwd();
+// Single authority for the "current project root" selection, shared across the
+// app/window, git, workspace-search, workspace-instructions, and session-entry
+// IPC surfaces. botIncoming, automation cron runs, and quick-chat read the
+// current selection through the thin `resolveCurrentProjectRoot` adapter below.
+const projectRootController = createProjectRootController({
+  lastProjectPathFile: join(workspaceRoot, 'last-project-path.json'),
+  fallbackRoots: () => [process.cwd(), app.getAppPath()],
+});
+const resolveCurrentProjectRoot: () => Promise<string> = () => projectRootController.current();
 const botRegistry = new BotRegistry({
   onIncomingMessage: (message: BotIncomingMessage) => {
     // Only log incoming bot messages in dev — production stdout leaking
@@ -874,22 +826,6 @@ async function resolveToolArtifactSourcePath(cwd: string, sourcePath: string): P
     return null;
   }
   return isInsideOrSamePath(root, target) ? target : null;
-}
-
-/**
- * Sanitize a single path segment for use under `screenshots/`. Allows
- * only `[a-zA-Z0-9._-]`; rejects everything else (slashes, `..`, NUL,
- * UTF-8 letters). Returns null when the input is empty after sanitization
- * so the capture IPC can fail-closed rather than write to an attacker-
- * controlled relative path.
- */
-function sanitizeSegment(value: unknown): string | null {
-  if (typeof value !== 'string') return null;
-  const trimmed = value.trim();
-  if (trimmed.length === 0 || trimmed.length > 128) return null;
-  if (!/^[a-zA-Z0-9._-]+$/.test(trimmed)) return null;
-  if (trimmed === '.' || trimmed === '..') return null;
-  return trimmed;
 }
 
 function isInsideOrSamePath(root: string, target: string): boolean {
@@ -1003,18 +939,6 @@ backends.register('ai-sdk', async (ctx) => {
   });
 });
 
-async function tryWeChatQrResult<T>(fn: () => Promise<T>, errorCode: string): Promise<Result<T>> {
-  try {
-    return ok(await fn());
-  } catch (error) {
-    return err(errorCode, weChatQrFailureMessage(error));
-  }
-}
-
-function weChatQrFailureMessage(error: unknown): string {
-  return generalizedErrorMessageChinese(error, '微信扫码登录暂时不可用，请稍后重试。');
-}
-
 backends.register('fake', (ctx) =>
   new FakeBackend({ sessionId: ctx.sessionId, header: ctx.header, store: ctx.store, appendMessage: ctx.appendMessage }),
 );
@@ -1090,411 +1014,51 @@ const onboardingService = createOnboardingService(
   }),
 );
 
-function workspaceInstructionOpenFailureCopy(reason: WorkspaceInstructionOpenFailureReason | 'open-failed'): string {
-  switch (reason) {
-    case 'unknown-file':
-      return '只能打开 AGENTS.md / CLAUDE.md / GEMINI.md。';
-    case 'missing':
-      return '项目指令文件不存在。';
-    case 'blocked':
-      return '项目指令文件不在当前工作区范围内。';
-    case 'not-a-file':
-      return '项目指令路径不是普通文件。';
-    case 'open-failed':
-      return '系统未能打开这个文件。';
-  }
-}
-
-function workspaceInstructionCreateFailureCopy(reason: WorkspaceInstructionCreateFailureReason): string {
-  switch (reason) {
-    case 'unknown-file':
-      return '只能创建 AGENTS.md / CLAUDE.md / GEMINI.md。';
-    case 'exists':
-      return '项目指令文件已经存在。';
-    case 'blocked':
-      return '当前工作区路径不可写或不在允许范围内。';
-    case 'write-failed':
-      return '写入项目指令文件失败。';
-  }
-}
-
-
-
-function proxyTestFailureMessage(result: TestProxyResult): string {
-  const raw = redactSecrets(result.error ?? '').trim();
-  const lower = raw.toLowerCase();
-  if (lower.includes('proxy disabled')) return '代理未启用，请先打开代理开关。';
-  if (lower.includes('proxy host/port required')) return '请填写代理服务器地址和端口后再测试。';
-  if (lower.includes('proxy test timeout') || lower.includes('timeout')) return '代理测试超时，请检查代理服务是否可达。';
-  if (result.status) return `代理测试返回 HTTP ${result.status}，请检查代理服务或测试地址。`;
-  const classified = generalizedErrorMessageChinese(raw, '');
-  if (classified) return classified;
-  if (raw && /[\u4E00-\u9FFF]/.test(raw)) return raw;
-  return '代理不可达，请检查代理服务器地址、端口或认证信息。';
-}
-
 function registerIpc(): void {
-  const LAST_PROJECT_PATH_FILE = join(workspaceRoot, 'last-project-path.json');
+  const currentProjectRoot = resolveCurrentProjectRoot;
 
-  let selectedProjectRoot: string | null = null;
-
-  async function loadPersistedProjectRoot(): Promise<string | null> {
-    try {
-      const raw = await readFile(LAST_PROJECT_PATH_FILE, 'utf8');
-      const parsed = JSON.parse(raw) as Record<string, unknown>;
-      if (typeof parsed.projectPath === 'string' && parsed.projectPath) {
-        await stat(parsed.projectPath);
-        return await resolveProjectRoot([parsed.projectPath]);
-      }
-    } catch {
-      // File missing, invalid, or points at a deleted directory.
-    }
-    return null;
-  }
-  const persistedProjectRootPromise = loadPersistedProjectRoot();
-
-  async function saveLastProjectPath(projectPath: string): Promise<void> {
-    try {
-      await writeFile(LAST_PROJECT_PATH_FILE, JSON.stringify({ projectPath }), 'utf8');
-    } catch {
-      // Best-effort; failure should not block the selection.
-    }
-  }
-
-  async function currentProjectRoot(): Promise<string> {
-    if (selectedProjectRoot) return selectedProjectRoot;
-    const persistedProjectRoot = await persistedProjectRootPromise;
-    if (persistedProjectRoot) {
-      selectedProjectRoot = persistedProjectRoot;
-      return persistedProjectRoot;
-    }
-    return resolveProjectRoot([process.cwd(), app.getAppPath()]);
-  }
-  resolveCurrentProjectRoot = currentProjectRoot;
-
-  async function resolveExplicitProjectRoot(projectPath: unknown): Promise<
-    | { ok: true; projectPath: string }
-    | { ok: false; reason: 'invalid-path' | 'not-found' }
-  > {
-    if (typeof projectPath !== 'string' || !projectPath) {
-      return { ok: false, reason: 'invalid-path' };
-    }
-    try {
-      await stat(projectPath);
-    } catch {
-      return { ok: false, reason: 'not-found' };
-    }
-    return { ok: true, projectPath: await resolveProjectRoot([projectPath]) };
-  }
-
-  ipcMain.handle('window:setTitlebarControlsVisible', (event, visible: unknown): void => {
-    mainWindowController.setTitlebarControlsVisible(event.sender, visible);
+  registerAppIpc({
+    mainWindowController,
+    projectRoot: projectRootController,
+    workspaceRoot,
+    buildInfo,
+    visualSmokeFixture,
   });
-  // PR-SHOW-AFTER-FIRST-COMMIT: the renderer signals its first React commit so
-  // the hidden window (main-window.ts show: false) is revealed only once real
-  // content can paint. Idempotent + visual-smoke-safe inside the controller.
-  ipcMain.handle('window:notifyRendererReady', (): void => {
-    mainWindowController.notifyRendererReady();
-  });
-  ipcMain.handle('window:setThemeSource', (event, themePref: unknown): void => {
-    mainWindowController.setThemeSource(event.sender, themePref);
-  });
-  // PR-WINDOW-TITLEBAR-0: re-sync the native titleBarOverlay color when the
-  // renderer resolves a new light/dark mode or palette. No-op outside Windows.
-  ipcMain.handle('window:setTitleBarOverlayTheme', (event, theme: unknown): void => {
-    mainWindowController.setTitleBarOverlayTheme(event.sender, theme);
-  });
-  ipcMain.handle('app:info', async () => {
-    const projectPath = await currentProjectRoot();
-    return {
-      appVersion: app.getVersion(),
-      electronVersion: process.versions.electron ?? '',
-      nodeVersion: process.versions.node ?? '',
-      chromeVersion: process.versions.chrome ?? '',
-      platform: process.platform,
-      arch: osArch(),
-      osRelease: osRelease(),
-      workspacePath: workspaceRoot,
-      projectPath,
-      projectGit: await resolveProjectGitInfo(projectPath),
-      buildMode: buildInfo.mode,
-      buildCommit: buildInfo.commit,
-    };
-  });
-  ipcMain.handle('app:openPath', async (_event, key: string): Promise<OpenPathResult> => {
-    const resolved = await resolveOpenPath({ key, workspaceRoot, projectRoot: await currentProjectRoot() });
-    if (!resolved.ok) return resolved;
-    const error = await shell.openPath(resolved.path);
-    if (error) return { ok: false, reason: 'open-failed' };
-    return { ok: true, opened: resolved.key };
-  });
-  ipcMain.handle(
-    'app:selectProjectDirectory',
-    async (): Promise<
-      | { ok: true; projectPath: string; projectGit: Awaited<ReturnType<typeof resolveProjectGitInfo>> }
-      | { ok: false; reason: 'cancelled' | 'missing-selection' }
-    > => {
-      const result = await mainWindowController.showOpenDialog({
-        title: '选择工作目录',
-        properties: ['openDirectory'],
-      });
-      const selectedPath = result.filePaths[0];
-      if (result.canceled) return { ok: false, reason: 'cancelled' };
-      if (!selectedPath) return { ok: false, reason: 'missing-selection' };
-      const projectPath = await resolveProjectRoot([selectedPath]);
-      selectedProjectRoot = projectPath;
-      void saveLastProjectPath(projectPath);
-      return {
-        ok: true,
-        projectPath,
-        projectGit: await resolveProjectGitInfo(projectPath),
-      };
-    },
-  );
-  ipcMain.handle(
-    'app:selectProjectRoot',
-    async (_event, projectPath: unknown): Promise<
-      | { ok: true; projectPath: string; projectGit: Awaited<ReturnType<typeof resolveProjectGitInfo>> }
-      | { ok: false; reason: 'invalid-path' | 'not-found' }
-    > => {
-      const explicitRoot = await resolveExplicitProjectRoot(projectPath);
-      if (!explicitRoot.ok) return explicitRoot;
-      const resolved = explicitRoot.projectPath;
-      selectedProjectRoot = resolved;
-      void saveLastProjectPath(resolved);
-      return {
-        ok: true,
-        projectPath: resolved,
-        projectGit: await resolveProjectGitInfo(resolved),
-      };
-    },
-  );
-  ipcMain.handle(
-    'app:resolveProjectGitInfo',
-    async (
-      _event,
-      projectPath: unknown,
-    ): Promise<
-      | { ok: true; projectPath: string; projectGit: Awaited<ReturnType<typeof resolveProjectGitInfo>> }
-      | { ok: false; reason: 'invalid-path' | 'not-found' }
-    > => {
-      if (projectPath !== undefined) {
-        const explicitRoot = await resolveExplicitProjectRoot(projectPath);
-        if (!explicitRoot.ok) return explicitRoot;
-        const resolved = explicitRoot.projectPath;
-        return { ok: true, projectPath: resolved, projectGit: await resolveProjectGitInfo(resolved) };
-      }
-      const resolved = await currentProjectRoot();
-      return { ok: true, projectPath: resolved, projectGit: await resolveProjectGitInfo(resolved) };
-    },
-  );
-  ipcMain.handle('app:listGitBranches', async () => {
-    const projectPath = await currentProjectRoot();
-    return listLocalBranches(projectPath);
-  });
-  ipcMain.handle(
-    'app:checkoutGitBranch',
-    async (_event, branch: unknown): Promise<{ ok: boolean; branch?: string; reason?: string; message?: string }> => {
-      if (typeof branch !== 'string' || !branch) {
-        return { ok: false, reason: 'failed', message: '无效的分支名' };
-      }
-      const projectPath = await currentProjectRoot();
-      return checkoutBranch(projectPath, branch);
-    },
-  );
-  // Composer `@` mention popup: list workspace files under the same project
-  // root that app:info reports. Git repos honor .gitignore + untracked via
-  // `git ls-files`; other trees fall back to a bounded walk. See
-  // workspace-file-search.ts.
-  ipcMain.handle(
-    'workspace:searchFiles',
-    async (_event, input: unknown) => {
-      const request = (input ?? {}) as { query?: unknown; limit?: unknown };
-      const projectPath = await currentProjectRoot();
-      return searchWorkspaceFiles(projectPath, { query: request.query, limit: request.limit });
-    },
-  );
   registerMemoryIpc({ localMemory });
   registerConfigIpc({ connectionStore, settingsStore, credentialStore, workspaceRoot });
   registerNotificationsIpc({ settingsStore, mainWindowController, e2e: isE2e });
-  ipcMain.handle('workspaceInstructions:getState', async () => getWorkspaceInstructionsState(await currentProjectRoot()));
-  ipcMain.handle(
-    'workspaceInstructions:openFile',
-    async (_event, file: unknown): Promise<{ ok: true } | { ok: false; message: string }> => {
-      const resolved = await resolveWorkspaceInstructionFileForOpen(await currentProjectRoot(), typeof file === 'string' ? file : '');
-      if (!resolved.ok) return { ok: false, message: workspaceInstructionOpenFailureCopy(resolved.reason) };
-      const error = await shell.openPath(resolved.path);
-      return error ? { ok: false, message: workspaceInstructionOpenFailureCopy('open-failed') } : { ok: true };
-    },
-  );
-  ipcMain.handle(
-    'workspaceInstructions:createFile',
-    async (_event, file: unknown): Promise<{ ok: true } | { ok: false; message: string }> => {
-      const created = await createWorkspaceInstructionFile(await currentProjectRoot(), typeof file === 'string' ? file : '');
-      if (!created.ok) return { ok: false, message: workspaceInstructionCreateFailureCopy(created.reason) };
-      return { ok: true };
-    },
-  );
+  registerWorkspaceInstructionsIpc({ getCurrentProjectRoot: currentProjectRoot });
   registerWorkspaceResourcesIpc({
     workspaceRoot,
     artifactStore,
     mainWindowController,
     sendToRenderer: safeSendToRenderer,
   });
-  ipcMain.handle('visualSmoke:getState', () => getVisualSmokeState(visualSmokeFixture));
-  /**
-   * PR-IR-01 screenshot capture (dev/test-only).
-   *
-   * Available only when `MAKA_VISUAL_SMOKE_FIXTURE` is set — refuses
-   * otherwise so real users / packaged builds can't be coerced into
-   * dumping the renderer to disk. The capture script
-   * (`scripts/capture-screenshots.mjs`) drives this IPC after the
-   * fixture finishes settling.
-   *
-   * Returns the absolute path of the written file or a structured
-   * failure reason. The renderer never sees absolute paths (per the
-   * filesystem-boundary contract); the script reads the result back
-   * over IPC because it owns the screenshot directory.
-   */
-  ipcMain.handle(
-    'visualSmoke:capture',
-    async (
-      _event,
-      input: { scenario: string; variant: string },
-    ): Promise<
-      | { ok: true; path: string }
-      | { ok: false; reason: 'not_in_fixture_mode' | 'invalid_input' | 'capture_failed' | 'write_failed' }
-    > => {
-      if (!visualSmokeFixture) return { ok: false, reason: 'not_in_fixture_mode' };
-      const scenario = sanitizeSegment(input?.scenario);
-      const variant = sanitizeSegment(input?.variant);
-      if (!scenario || !variant) return { ok: false, reason: 'invalid_input' };
-      let image: Electron.NativeImage;
-      try {
-        const capture = await mainWindowController.capturePage();
-        if (!capture) return { ok: false, reason: 'capture_failed' };
-        image = capture;
-      } catch {
-        return { ok: false, reason: 'capture_failed' };
-      }
-      const dir = join(workspaceRoot, 'screenshots', scenario);
-      try {
-        await mkdir(dir, { recursive: true });
-      } catch {
-        return { ok: false, reason: 'write_failed' };
-      }
-      const filePath = join(dir, `${variant}.png`);
-      try {
-        const { writeFile } = await import('node:fs/promises');
-        await writeFile(filePath, image.toPNG());
-      } catch {
-        return { ok: false, reason: 'write_failed' };
-      }
-      // Deterministic stdout marker so the driver script
-      // (`scripts/capture-screenshots.mjs`) can match on the line and
-      // know the capture completed without polling the filesystem.
-      // The line is single-token whitespace-separated so it's easy to
-      // parse by regex.
-      console.log(`[visual-smoke] captured scenario=${scenario} variant=${variant} path=${filePath}`);
-      return { ok: true, path: filePath };
-    },
-  );
+  registerWorkspaceSearchIpc({ getCurrentProjectRoot: currentProjectRoot });
+  registerGitIpc({ getCurrentProjectRoot: currentProjectRoot });
   registerPlanReminderIpc({ planReminders, getWorkspacePrivacyContext });
-  ipcMain.handle('shell-runs:list', (_event, sessionId: string) => runtime.listShellRunUpdates(sessionId));
-  ipcMain.handle('tasks:list', async (_event, sessionId: string) => {
-    const tasks = await taskLedgerStore.list(sessionId, {
-      includeTerminal: true,
-      includeArchived: false,
-      classifyResumeTrust: true,
-      ...(visualSmokeFixture ? { now: getVisualSmokeState(visualSmokeFixture)?.now ?? Date.now() } : {}),
-    });
-    return tasks.map(sanitizeTaskLedgerTask);
+  registerSessionsIpc({
+    runtime,
+    store,
+    taskLedgerStore,
+    goalManager: goalWiring.manager,
+    automationManager: automationWiring.manager,
+    computerUseOverlay,
+    computerUseTools,
+    artifactStore,
+    attachmentApprovals,
+    settingsStore,
+    connectionStore,
+    mainWindowController,
+    visualSmokeFixture,
+    emitSessionsChanged,
+    ensureSessionCanSend,
+    getReadyConnection,
+    streamEvents,
+    getCurrentProjectRoot: currentProjectRoot,
+    getWorkspacePrivacyContext,
+    canCreateFakeSession: canCreateFakeSessionFromRenderer,
   });
-  ipcMain.handle('sessions:list', (_event, filter?: SessionListFilter) => runtime.listSessions(filter));
-  ipcMain.handle('sessions:create', async (_event, input?: Partial<CreateSessionInput>) => {
-    const cwd = input?.cwd ?? (await currentProjectRoot());
-    if (input?.backend === 'fake') {
-      if (!canCreateFakeSessionFromRenderer()) {
-        throw new Error('FakeBackend sessions are only available in development.');
-      }
-      const session = await runtime.createSession({
-        cwd,
-        backend: 'fake',
-        llmConnectionSlug: input.llmConnectionSlug ?? 'fake',
-        model: input.model ?? 'fake-model',
-        permissionMode: input.permissionMode ?? (await resolveDefaultPermissionMode(() => settingsStore.get())),
-        name: input.name ?? 'New Chat',
-        labels: input.labels,
-      });
-      emitSessionsChanged('created', session.id);
-      return session;
-    }
-
-    const requestedSlug = input?.llmConnectionSlug ?? (await connectionStore.getDefault());
-    const { connection, model } = await getReadyConnection(requestedSlug, input?.model);
-    const thinkingLevel = normalizeSupportedSessionThinkingLevel(input?.thinkingLevel, connection.providerType, model);
-
-    const session = await runtime.createSession({
-      cwd,
-      backend: 'ai-sdk',
-      llmConnectionSlug: connection.slug,
-      model,
-      ...(thinkingLevel !== undefined ? { thinkingLevel } : {}),
-      permissionMode: input?.permissionMode ?? (await resolveDefaultPermissionMode(() => settingsStore.get())),
-      name: input?.name ?? 'New Chat',
-      labels: input?.labels,
-    });
-    emitSessionsChanged('created', session.id);
-    return session;
-  });
-  ipcMain.handle('sessions:readMessages', async (_event, sessionId: string) => {
-    if (visualSmokeFixture) return store.readMessages(sessionId);
-    let messages: StoredMessage[];
-    try {
-      messages = await runtime.getMessages(sessionId);
-    } catch (error) {
-      throw new Error(sessionReadMessagesFailureMessage(error));
-    }
-    try {
-      await runtime.markSessionRead(sessionId, latestStoredMessageTs(messages));
-    } catch {
-      // Reading the content already succeeded. Leave the persisted unread
-      // state for a later refresh instead of turning this into a load error.
-    }
-    return messages;
-  });
-  ipcMain.handle('sessions:listTurns', (_event, sessionId: string) => runtime.listTurns(sessionId));
-  // Goal kill-switch surface: the renderer reads the active goal to badge a
-  // session running an autonomous loop, and clears it to stop the loop. `get`
-  // returns null when no goal is set; `clear` settles it (continuation stops
-  // after the current turn). Both are pure local state, so no permission gate.
-  ipcMain.handle('goal:get', (_event, sessionId: string) => goalWiring.manager.get(sessionId) ?? null);
-  ipcMain.handle('goal:clear', (_event, sessionId: string) => {
-    goalWiring.manager.clear(sessionId);
-  });
-  // PR-SEARCH-2: local thread search. Renderer-facing channel; the pure
-  // helper in `./search/thread-search.ts` enforces all gates (G1 snippet
-  // redaction, G2 fake-backend exclude, G4 caps, G5 case-fold + NFC,
-  // G9 tool_result scan cap, G10 system/meta exclusion). The helper
-  // receives the runtime via DI so unit tests stay Electron-agnostic.
-  // We deliberately do NOT log the request body — query text never enters
-  // telemetry.
-  // ===========================================================
-  // PR-OAUTH-SUBSCRIPTION-0: Claude subscription OAuth IPC.
-  // All handlers return either `SubscriptionAccountState` or
-  // `SubscriptionActionResult` — never raw tokens (xuan G-X3).
-  //
-  // kenji `1da909d5` blocking concern: Anthropic does not permit
-  // third-party developers to offer Claude.ai login on behalf of
-  // users. Until product/legal sign-off, the entire feature is
-  // gated behind `MAKA_CLAUDE_SUBSCRIPTION_EXPERIMENTAL=1`. The
-  // Settings UI also hides the card; this guard is the second line
-  // of defense (a DevTools-triggered call to `window.maka` still
-  // hits the experimental gate).
-  // ===========================================================
-  // kenji `45b31e16`: use the dedicated `experimental_disabled`
-  // reason so the user-visible state is clearly "this feature is
-  // not enabled by Maka" — NOT "Anthropic rejected my account".
   registerSubscriptionIpc({
     connectionStore,
     claudeSubscription,
@@ -1509,205 +1073,8 @@ function registerIpc(): void {
     syncGitHubCopilotConnection,
     emitConnectionListChanged,
   });
-
   registerWebSearchIpc({ settingsStore, getWorkspacePrivacyContext });
-
-  ipcMain.handle('search:thread', async (_event, request: unknown) => {
-    // PR-SEARCH-2 review fixup (@xuan `2f1aba55`): pass `unknown`
-    // through to the helper, which runs an object-shape guard and
-    // returns an `invalid_query` error envelope for null / non-object
-    // / missing-field payloads. Never throws across the IPC boundary.
-    //
-    // PR-SEARCH-2.5 (@xuan `2c55b975`): wire `getPrivacyContext` to
-    // the main-authority workspace privacy state.
-    //
-    // This is the main-owned workspace privacy source, not a renderer
-    // self-attestation. The helper validates whatever shape is returned
-    // via `validateWorkspacePrivacyContext`, so a future drift in
-    // authority source is automatically fail-closed.
-    return runThreadSearch(request, {
-      listSessions: () => runtime.listSessions(),
-      readMessages: (sessionId: string) => runtime.getMessages(sessionId),
-      getPrivacyContext: getWorkspacePrivacyContext,
-    });
-  });
-  ipcMain.handle('sessions:stop', async (_event, sessionId: string, input?: { source?: 'stop_button' }) => {
-    computerUseOverlay.clearForSession(sessionId);
-    computerUseTools.clearSession(sessionId);
-    await runtime.stopSession(sessionId, normalizeStopSessionInput(input));
-    emitSessionsChanged('status-change', sessionId);
-    emitSessionsChanged('turn-status-change', sessionId);
-    emitSessionsChanged('message-appended', sessionId);
-  });
-  ipcMain.handle('sessions:respondToPermission', (_event, sessionId: string, response) =>
-    runtime.respondToPermission(sessionId, normalizePermissionResponse(response)),
-  );
-  ipcMain.handle('sessions:respondToUserQuestion', (_event, sessionId: string, response) =>
-    runtime.respondToUserQuestion(sessionId, normalizeUserQuestionResponse(response)),
-  );
-  ipcMain.handle('sessions:send', async (event, sessionId: string, command: unknown) => {
-    const sendCommand = normalizeSessionSendCommand(command);
-    if (!sendCommand) return;
-    const { turnId, attachments } = await resolveSessionSend({
-      sessionId,
-      senderId: event.sender.id,
-      command: sendCommand,
-      ensureCanSend: ensureSessionCanSend,
-      readHeader: (id) => store.readHeader(id),
-      approvals: attachmentApprovals,
-      stat: async (path) => ({ size: (await stat(path)).size }),
-      artifactStore,
-      resizeImage: resizeImageForAttachment,
-    });
-    const iterator = runtime.sendMessage(sessionId, {
-      turnId,
-      text: sendCommand.text,
-      ...(attachments.length > 0 ? { attachments } : {}),
-    });
-    void streamEvents(sessionId, iterator, turnId);
-    return { turnId, attachments };
-  });
-  ipcMain.handle(
-    'attachments:pickFiles',
-    async (event): Promise<
-      | { ok: true; files: { approvalId: string; name: string; mimeType?: string; size: number }[] }
-      | { ok: false; reason: 'cancelled' }
-    > => {
-      const result = await mainWindowController.showOpenDialog({
-        title: '添加附件',
-        properties: ['openFile', 'multiSelections'],
-      });
-      if (result.canceled || !result.filePaths[0]) return { ok: false, reason: 'cancelled' };
-      const chosen = await Promise.all(
-        result.filePaths.map(async (path) => ({ path, name: basename(path), size: (await stat(path)).size })),
-      );
-      // Paths stay in main; the renderer only gets one-shot opaque tokens.
-      return { ok: true, files: attachmentApprovals.issueApprovals(event.sender.id, chosen) };
-    },
-  );
-  ipcMain.handle(
-    'attachments:readBytes',
-    async (_event, sessionId: string, relativePath: string): Promise<
-      | { ok: true; base64: string; mimeType: string }
-      | { ok: false; reason: string }
-    > => {
-      // Session-scoped read: only attachments filed under this session.
-      const record = await artifactStore.get(relativePath).catch(() => null);
-      if (!record || record.sessionId !== sessionId) return { ok: false, reason: 'not_found' };
-      const result = await artifactStore.readBinary(relativePath);
-      if (!result.ok) return result;
-      return { ok: true, base64: result.base64, mimeType: result.mimeType };
-    },
-  );
-  ipcMain.handle('sessions:compact', async (_event, sessionId: string) => {
-    await ensureSessionCanSend(sessionId);
-    const turnId = randomUUID();
-    void streamEvents(sessionId, runtime.compactSession(sessionId, { turnId }), turnId);
-  });
-  ipcMain.handle('sessions:regenerateTurn', async (_event, sessionId: string, input: unknown) => {
-    await ensureSessionCanSend(sessionId);
-    const normalized = normalizeRegenerateTurnInput(input);
-    const turnId = normalized.turnId ?? randomUUID();
-    void streamEvents(sessionId, runtime.regenerateTurn(sessionId, { ...normalized, turnId }), turnId);
-  });
-  ipcMain.handle('sessions:branchFromTurn', async (_event, sessionId: string, input: unknown) => {
-    const session = await runtime.branchFromTurn(sessionId, normalizeBranchFromTurnInput(input));
-    emitSessionsChanged('created', session.id);
-    return session;
-  });
-  ipcMain.handle('sessions:archive', async (_event, sessionId: string) => {
-    computerUseOverlay.clearForSession(sessionId);
-    computerUseTools.clearSession(sessionId);
-    await runtime.archive(sessionId);
-    // An archived conversation is no longer shown: drop its browser connection
-    // and view so it does not keep a live Chromium page in the background.
-    await releaseBrowserSession(sessionId);
-    // Stop any autonomous loops tied to the session (goal + polling heartbeats).
-    goalWiring.manager.remove(sessionId);
-    automationWiring.manager.removeAllForSession(sessionId);
-    emitSessionsChanged('archived', sessionId);
-  });
-  ipcMain.handle('sessions:unarchive', async (_event, sessionId: string) => {
-    await runtime.unarchive(sessionId);
-    emitSessionsChanged('updated', sessionId);
-  });
-  ipcMain.handle('sessions:setFlagged', async (_event, sessionId: string, isFlagged: boolean) => {
-    await runtime.setFlagged(sessionId, isFlagged);
-    emitSessionsChanged('pinned', sessionId);
-  });
-  ipcMain.handle('sessions:rename', async (_event, sessionId: string, name: string) => {
-    await runtime.renameSession(sessionId, name);
-    emitSessionsChanged('renamed', sessionId);
-  });
-  ipcMain.handle('sessions:setPermissionMode', (_event, sessionId: string, mode: unknown) => {
-    if (!isPermissionMode(mode)) {
-      throw new Error(`Invalid permission mode: ${String(mode)}`);
-    }
-    return runtime.setPermissionMode(sessionId, mode).then((session) => {
-      emitSessionsChanged('mode-change', sessionId);
-      return session;
-    });
-  });
-  ipcMain.handle('sessions:setModel', async (_event, sessionId: string, input: unknown) => {
-    const { llmConnectionSlug, model } = normalizeSessionModelSelection(input);
-    const header = await store.readHeader(sessionId);
-    if (header.status === 'running') {
-      throw new Error('当前对话正在运行，等结束后再切换模型。');
-    }
-    if (header.status === 'waiting_for_user') {
-      throw new Error('当前有工具调用正在等待确认，处理后再切换模型。');
-    }
-    const ready = await getReadyConnection(llmConnectionSlug, model);
-    const next = await runtime.updateSession(sessionId, {
-      backend: 'ai-sdk',
-      llmConnectionSlug: ready.connection.slug,
-      model: ready.model,
-      // Switching model clears the per-model thinking variant (see model-thinking.ts).
-      thinkingLevel: undefined,
-      connectionLocked: true,
-      status: 'active',
-      blockedReason: undefined,
-      statusUpdatedAt: Date.now(),
-    });
-    emitSessionsChanged('updated', sessionId, {
-      connectionSlug: ready.connection.slug,
-      modelId: ready.model,
-    });
-    return next;
-  });
-  ipcMain.handle('sessions:setThinkingLevel', async (_event, sessionId: string, input: unknown) => {
-    const header = await store.readHeader(sessionId);
-    if (header.status === 'running') {
-      throw new Error('当前对话正在运行，等结束后再切换思考级别。');
-    }
-    if (header.status === 'waiting_for_user') {
-      throw new Error('当前有工具调用正在等待确认，处理后再切换思考级别。');
-    }
-    const connection = await connectionStore.get(header.llmConnectionSlug);
-    if (!connection) {
-      throw new Error(`Unknown connection: ${header.llmConnectionSlug}`);
-    }
-    const nextThinkingLevel = normalizeSupportedSessionThinkingLevel(input, connection.providerType, header.model);
-    const next = await runtime.updateSession(sessionId, nextThinkingLevel === undefined ? { thinkingLevel: undefined } : { thinkingLevel: nextThinkingLevel });
-    emitSessionsChanged('updated', sessionId);
-    return next;
-  });
-  ipcMain.handle('sessions:remove', async (_event, sessionId: string) => {
-    computerUseOverlay.clearForSession(sessionId);
-    computerUseTools.clearSession(sessionId);
-    await runtime.remove(sessionId);
-    // Drop the conversation's browser connection and destroy its view (no-op
-    // if it never opened one). releaseBrowserSession disposes the view via the
-    // host, covering both agent-driven and hand-opened views.
-    await releaseBrowserSession(sessionId);
-    // Stop any autonomous loops tied to the session (goal + polling heartbeats).
-    goalWiring.manager.remove(sessionId);
-    automationWiring.manager.removeAllForSession(sessionId);
-    emitSessionsChanged('deleted', sessionId);
-  });
-
   registerBrowserIpc({ mainWindowController });
-
   registerConnectionsIpc({
     connectionStore,
     credentialStore,
@@ -1716,214 +1083,31 @@ function registerIpc(): void {
     hasConnectionSecret,
     emitConnectionListChanged,
   });
-
-  // PR110b: Onboarding snapshot + milestone IPCs. Renderer polls via
-  // these on app load and whenever `sessions:changed` /
-  // `connections:changed` / settings change events fire. No push from
-  // main.
-  ipcMain.handle('onboarding:getSnapshot', async () => onboardingService.getSnapshot());
-  ipcMain.handle('onboarding:setMilestone', async (_event, id: unknown, status: unknown) => {
-    // Service throws INVALID_MILESTONE_ID / INVALID_MILESTONE_STATUS
-    // for bad inputs; let the error propagate so the renderer sees
-    // it as a typed reject rather than silently swallowing.
-    return onboardingService.setMilestone(id, status);
+  registerOnboardingIpc({ onboardingService });
+  registerSessionEntryIpc({
+    runtime,
+    getReadyConnection,
+    getCurrentProjectRoot: currentProjectRoot,
+    getOnboardingState: async () => (await onboardingService.getSnapshot()).state,
+    emitSessionsChanged,
+    ensureSessionCanSend,
+    streamEvents,
+    quickChatStart: (input) => handleQuickChatStart(input, currentProjectRoot),
   });
-  ipcMain.handle('onboarding:clearMilestone', async (_event, id: unknown) => {
-    return onboardingService.clearMilestone(id);
+  registerPermissionsIpc({
+    settingsStore,
+    connectionStore,
+    telemetryRepo,
+    botRegistry,
+    getComputerUseCapabilityInput: computerUseCapabilityInput,
   });
-  // PR110b: Quick Chat entry. Input shape is intentionally minimal —
-  // `{ prompt?: string }` — to keep readiness gating airtight. Override
-  // surfaces (connectionSlug / model) will land in PR110c/d when the
-  // model-picker UI is ready.
-  ipcMain.handle('quickChat:start', async (_event, input: unknown) => {
-    return handleQuickChatStart(input, currentProjectRoot);
+  registerSettingsIpc({
+    settingsStore,
+    botRegistry,
+    normalizeSettingsPatch,
+    applySettingsRuntimeEffects,
   });
-
-  // Expert teams: list the built-in teams and start a labeled team session.
-  // A team session is a normal session tagged `mode:expert-team:<teamId>`; the
-  // label activates the lead persona + expert_dispatch tool (see the backend
-  // factory). The lead runs read-only (explore) and dispatches read-only members.
-  ipcMain.handle('expertTeam:list', async () => ({
-    teams: listExpertTeams().map((team) => ({
-      id: team.id,
-      name: team.name,
-      description: team.description,
-      members: team.members.map((member) => ({
-        id: member.id,
-        name: member.name,
-        description: member.description,
-        ...(member.whenToUse ? { whenToUse: member.whenToUse } : {}),
-      })),
-    })),
-  }));
-  ipcMain.handle('expertTeam:start', async (_event, input: unknown) => {
-    return runExpertTeamStart(input, {
-      isKnownTeam: (teamId) => getExpertTeam(teamId) !== undefined,
-      getOnboardingState: async () => (await onboardingService.getSnapshot()).state,
-      createSession: async ({ teamId, defaultConnectionSlug, defaultModel }) => {
-        const ready = await getReadyConnection(defaultConnectionSlug, defaultModel);
-        const team = getExpertTeam(teamId);
-        return runtime.createSession({
-          cwd: await currentProjectRoot(),
-          backend: 'ai-sdk',
-          llmConnectionSlug: ready.connection.slug,
-          model: ready.model,
-          // Shipped teams are read-only review crews: the lead reads + dispatches
-          // read-only members, so the whole session stays in explore mode.
-          permissionMode: 'explore',
-          name: team ? team.name : 'Expert Team',
-          labels: [expertTeamLabel(teamId)],
-        });
-      },
-      emitCreated: (sessionId) => emitSessionsChanged('created', sessionId),
-      ensureCanSend: (sessionId) => ensureSessionCanSend(sessionId),
-      sendFirstMessage: async (sessionId, text) => {
-        const turnId = randomUUID();
-        const iterator = runtime.sendMessage(sessionId, { turnId, text });
-        void streamEvents(sessionId, iterator, turnId);
-      },
-    });
-  });
-
-  ipcMain.handle('permissions:getSnapshot', () => buildPermissionSnapshot());
-  ipcMain.handle('permissions:openSystemSettings', async (_event, permId: unknown) => {
-    return openSystemPermissionPane(permId);
-  });
-  ipcMain.handle('permissions:requestAccess', async (_event, permId: unknown) => {
-    return requestPermissionAccess(permId);
-  });
-  ipcMain.handle('capabilities:getSnapshot', async () => {
-    const permissions = buildPermissionSnapshot();
-    const settings = await settingsStore.get();
-    const officeCliProbe = await probeOfficeCli({ now: permissions.checkedAt });
-    return buildCapabilitySnapshotCollection({
-      settings,
-      permissions,
-      botStatuses: botRegistry.allStatuses(),
-      officeCliProbe,
-      computerUse: computerUseCapabilityInput(),
-      now: permissions.checkedAt,
-    });
-  });
-  ipcMain.handle('health:getSnapshot', async () => {
-    const now = Date.now();
-    const permissions = buildPermissionSnapshot(now);
-    const settings = await settingsStore.get();
-    const officeCliProbe = await probeOfficeCli({ now });
-    const capabilitySnapshot = buildCapabilitySnapshotCollection({
-      settings,
-      permissions,
-      botStatuses: botRegistry.allStatuses(),
-      officeCliProbe,
-      computerUse: computerUseCapabilityInput(),
-      now,
-    });
-    const connections = await connectionStore.list();
-    const connectionSignals = connections.flatMap((connection) => [
-      healthSignalFromConnection(connection, now),
-      healthSignalFromConnectionRuntime(
-        connection,
-        telemetryRepo.latestLlmRuntimeProbe(connection.slug, connection.defaultModel),
-        now,
-      ),
-    ].filter((signal): signal is NonNullable<typeof signal> => Boolean(signal)));
-    return buildHealthSnapshot(now, [
-      ...connectionSignals,
-      ...capabilitySnapshot.capabilities.map(healthSignalFromCapability),
-    ]);
-  });
-
-  ipcMain.handle('settings:get', async () => maskAppSettings(await settingsStore.get()));
-  ipcMain.handle('settings:update', async (_event, patch: UpdateAppSettingsInput): Promise<UpdateAppSettingsResult> => {
-    const normalizedPatch = await normalizeSettingsPatch(patch);
-    const next = await settingsStore.update(normalizedPatch);
-    await applySettingsRuntimeEffects(next, patch);
-    return buildSettingsUpdateResult(next, patch);
-  });
-  ipcMain.handle('gateway:status', async () => openGateway.getStatus());
-  ipcMain.handle('settings:testNetworkProxy', async (_event, input: TestProxyInput = {}) => {
-    const started = Date.now();
-    const stored = toContractNetworkSettings((await settingsStore.get()).network).proxy;
-    const proxy = input.proxy?.password === SENSITIVE_PLACEHOLDER
-      ? { ...input.proxy, password: stored.password }
-      : input.proxy;
-    const testedProxy = proxy ?? stored;
-    const result = await testProxyConnection({ ...input, proxy }, stored);
-    const latencyMs = result.latencyMs ?? (Date.now() - started);
-    if (!result.ok) {
-      return {
-        ok: false,
-        message: proxyTestFailureMessage(result),
-        latencyMs,
-      } satisfies SettingsTestResult;
-    }
-    return {
-      ok: true,
-      message: result.ip
-        ? `代理配置有效：${testedProxy.type}://${testedProxy.host}:${testedProxy.port} · ${result.countryFlag ?? ''} ${result.ip}`.trim()
-        : `代理配置有效：${testedProxy.type}://${testedProxy.host}:${testedProxy.port}`,
-      latencyMs,
-      details: {
-        status: result.status,
-        ip: result.ip,
-        countryCode: result.countryCode,
-        countryFlag: result.countryFlag,
-        bypassList: testedProxy.bypassList,
-      },
-    } satisfies SettingsTestResult;
-  });
-  ipcMain.handle('settings:testBotChannel', async (_event, provider: BotProvider) => {
-    const settings = await settingsStore.get();
-    const result = await testRuntimeBotChannel(provider, settings.botChat.channels[provider]);
-    await settingsStore.update({
-      botChat: {
-        channels: {
-          [provider]: {
-            connected: result.ok,
-            readiness: result.ok ? 'credentials_valid' : 'configured',
-            readinessReason: result.ok ? undefined : botTestErrorMessage(provider, result.error),
-            readinessUpdatedAt: Date.now(),
-            lastTestAt: Date.now(),
-            lastError: result.ok ? undefined : botTestErrorMessage(provider, result.error),
-          },
-        },
-      },
-    });
-    const next = await settingsStore.get();
-    await applySettingsRuntimeEffects(next, { botChat: { channels: { [provider]: {} } } });
-    return toSettingsTestResult(provider, result);
-  });
-  ipcMain.handle('settings:bots:listStatuses', () =>
-    tryResult(async () => botRegistry.allStatuses(), 'BOTS_STATUS_FAILED'),
-  );
-  ipcMain.handle('settings:bots:restart', (_event, provider: BotProvider) =>
-    tryResult(async () => {
-      const settings = await settingsStore.get();
-      await botRegistry.applySettings(settings.botChat);
-      return botRegistry.getStatus(provider);
-    }, 'BOTS_RESTART_FAILED'),
-  );
-
-  // PR-BOT-WECHAT-QR-MODAL-0 (WAWQAQ msg `10ec1fbe`): WeChat ClawBot
-  // scan-login. Renderer triggers the QR fetch from the modal, then
-  // polls the status endpoint until 'confirmed' or 'expired'. Main
-  // process owns the actual HTTP calls so the renderer never sees
-  // raw response bodies.
-  ipcMain.handle('settings:bots:wechat:fetchQrcode', () =>
-    tryWeChatQrResult(async () => fetchWeChatQrcode(), 'WECHAT_QR_FETCH_FAILED'),
-  );
-  ipcMain.handle('settings:bots:wechat:pollQrcodeStatus', (_event, qrToken: unknown) =>
-    tryWeChatQrResult(async () => {
-      if (typeof qrToken !== 'string' || !qrToken) {
-        throw new Error('qrToken must be a non-empty string');
-      }
-      return pollWeChatQrcodeStatus(qrToken);
-    }, 'WECHAT_QR_STATUS_FAILED'),
-  );
-  ipcMain.handle('settings:bots:wechatQrCode', async () => {
-    const settings = await settingsStore.get();
-    return getWechatBridgeQrCode(settings.botChat.channels.wechat);
-  });
+  registerGatewayIpc({ openGateway });
   registerDailyReviewIpc({ dailyReview, dailyReviewArchiveStore, mainWindowController });
   registerUsageIpc({
     settingsStore,
@@ -1933,7 +1117,6 @@ function registerIpc(): void {
     },
     sendToRenderer: safeSendToRenderer,
   });
-
 }
 
 function canCreateFakeSessionFromRenderer(): boolean {
@@ -2076,14 +1259,6 @@ function isStatusChangingSessionEvent(event: SessionEvent): boolean {
     event.type === 'error';
 }
 
-function latestStoredMessageTs(messages: readonly StoredMessage[]): number | undefined {
-  let latest: number | undefined;
-  for (const message of messages) {
-    if (Number.isFinite(message.ts)) latest = latest === undefined ? message.ts : Math.max(latest, message.ts);
-  }
-  return latest;
-}
-
 function isTurnStatusChangingSessionEvent(event: SessionEvent): boolean {
   return event.type === 'complete' || event.type === 'abort' || event.type === 'error';
 }
@@ -2123,22 +1298,6 @@ const readyConnectionDeps = {
 
 function getReadyConnection(slug: string | null | undefined, model?: string) {
   return requireReadyConnection(slug, readyConnectionDeps, model);
-}
-
-function normalizeSupportedSessionThinkingLevel(
-  input: unknown,
-  providerType: ProviderType,
-  model: string,
-): ThinkingLevel | undefined {
-  const thinkingLevel = input === undefined || input === null ? undefined : input;
-  if (thinkingLevel === undefined) return undefined;
-  if (!isThinkingLevel(thinkingLevel)) {
-    throw new Error(`Invalid thinking level: ${String(input)}`);
-  }
-  if (!thinkingVariantsForModel(providerType, model).includes(thinkingLevel)) {
-    throw new Error(`当前模型不支持思考级别：${thinkingLevel}`);
-  }
-  return thinkingLevel;
 }
 
 /**
@@ -2215,22 +1374,6 @@ function emitSessionsChanged(
   if (extra?.connectionSlug) event.connectionSlug = extra.connectionSlug;
   if (extra?.modelId) event.modelId = extra.modelId;
   safeSendToRenderer('sessions:changed', event);
-}
-
-function normalizeSessionModelSelection(input: unknown): { llmConnectionSlug: string; model: string } {
-  if (!input || typeof input !== 'object') {
-    throw new Error('Invalid model selection');
-  }
-  const record = input as Record<string, unknown>;
-  const llmConnectionSlug = typeof record.llmConnectionSlug === 'string' ? record.llmConnectionSlug.trim() : '';
-  const model = typeof record.model === 'string' ? record.model.trim() : '';
-  if (!llmConnectionSlug) {
-    throw new Error('Missing model connection');
-  }
-  if (!model) {
-    throw new Error('Missing model');
-  }
-  return { llmConnectionSlug, model };
 }
 
 async function recoverInterruptedSessionsOnStartup(): Promise<void> {
