+++ scripts/check-console.mjs
@@ -39,6 +39,10 @@ const ALLOW = new Map([
     'dev-gated by VITE_DEV_SERVER_URL / NODE_ENV (PR100).',
   ],
   [
+    'apps/desktop/src/main/app-ipc-main.ts',
+    'visual-smoke capture marker is fixture-gated and stdout-parsed by capture tooling (moved from main.ts, #1084).',
+  ],
+  [
     'apps/desktop/src/main/daily-review-main.ts',
     'scheduler failures are main-process diagnostics and do not expose secrets.',
   ],
